+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add MC E2E test suite using `apptest-framework`, verifying all core Tempo components (distributor, ingester, querier, query-frontend, compactor, gateway) are ready on the management cluster.
+
 ## [0.14.0] - 2026-03-09
 
 ### Added